requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTX==1.3.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Serialization - PYDANTIC
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Password Hashing
bcrypt==4.1.2

# System Monitoring
psutil==5.9.6

# Environment Variables
python-dotenv==1.0.0

# Utilities
pytz==2023.3

# Testing
pytest==7.4.3
pytest-flask==1.3.0